--- backend/src/common/services/azure-ai/apis/apis.ts
@@ -1,2 +1,0 @@
-export { avatarVideoApi } from './avatar-video/avatar-video.js';
-export { textToSpeechApi } from './text-to-speech/text-to-speech.js';
