+++ backend/src/common/services/azure-ai/avatar-video/avatar-video.ts
@@ -1,7 +1,7 @@
 import { config } from '~/common/config/config.js';
 import { baseHttp } from '~/common/http/http.js';
 
-import { AvatarVideoApi } from './avatar-video-api.js';
+import { AvatarVideoApi } from './avatar-video-base.js';
 
 const avatarVideoApi = new AvatarVideoApi({
     http: baseHttp,
