+++ backend/src/common/services/azure-ai/avatar-video/enums/avatar-api-path.ts
