+++ backend/src/common/services/azure-ai/avatar-video/enums/enums.ts
