+++ backend/src/common/services/azure-ai/avatar-video/types/render-avatar-video-arguments.type.ts
