+++ backend/src/common/services/azure-ai/avatar-video/types/types.ts
