+++ backend/src/common/services/azure-ai/azure-ai.service.ts
@@ -13,8 +13,8 @@ import {
 import { type BaseConfig } from '~/common/config/base-config.package.js';
 
 import { type FileService } from '../file/file.service.js';
-import { type TextToSpeechApi } from './apis/text-to-speech/text-to-speech-api.js';
 import { DEFAULT_LANGUAGE } from './constants/constants.js';
+import { type TextToSpeechApi } from './text-to-speech/text-to-speech-base.js';
 import { type AzureGetVoicesResponseDto } from './types/types.js';
 
 class AzureAIService {
